refactor: move DEFAULT_PROMPT to const
