+++ src/const.ts
@@ -1,3 +1,7 @@
+export const DEFAULT_PROMPT: string =
+  `Classify the tab group base on the provided URL ({{tabURL}}) and title ({{tabTitle}}) into one of the categories: ` +
+  `{{types}}. Response with the category only, without any comments.`;
+
 export const DEFAULT_GROUP = [
   "Social",
   "Entertainment",
